WIP: Avoid undefined behavior when computing mipmap of small images
Flightgear may abort on startup when trying to compute the mipmap of
a small 1x1 image. An image so small has only one mipmap level which
results in an empty `mipmapOffsets` vector when we eventually call
`memcpy( data, image->data(), mipmapOffsets.front() )`. Calling
`front()` on an empty vector is undefined behavior, which in my
case triggers an abort with the following error:

    /usr/include/c++/11.1.0/bits/stl_vector.h:1123: std::vector<_Tp,
         _Alloc>::reference std::vector<_Tp, _Alloc>::front() [with
         _Tp = unsigned int; _Alloc = std::allocator<unsigned int>;
         std::vector<_Tp, _Alloc>::reference = unsigned int&]:
         Assertion '!this->empty()' failed.
    
    Thread 1 "fgfs" received signal SIGABRT, Aborted.

To work around this, detect when an image is too small to be be
reasonably mipmapped and generate a log message rather than
proceeding.